backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
flask_sql_alchemy==3.1.1
flask-migrate==4.1.0
flask-script==2.0.6
psycopg2-binary==2.9.10
